requirements.txt
colorama==0.4.6
distlib==0.3.8
filelock==3.13.1
pillow==10.2.0
platformdirs==4.1.0
virtualenv==20.25.0
